refactor: Extract converters into separate package
