+++ pkg/convert/slice_test.go
@@ -1,4 +1,4 @@
-package cslices
+package convert
 
 import "testing"
